[Refactor] prefer mixins
